backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
Pillow==11.2.1
openpyxl==3.1.5
reportlab==4.2.5
django-jazzmin==2.6.0